+++ common/src/main/java/com/food/common/foodCategory/utils/FoodCategoryFailureMessages.java
@@ -1,6 +1,6 @@
 package com.food.common.foodCategory.utils;
 
-import com.food.common.common.utils.ValidationFailureMessages;
+import com.food.common.utils.validation.ValidationFailureMessages;
 
 public class FoodCategoryFailureMessages extends ValidationFailureMessages {
     public static final String NOT_BLANK_NAME = "음식 카테고리명" + CANNOT_BE_BLANK;
