+++ common/src/main/java/com/food/common/user/domain/SocialAccount.java
@@ -1,6 +1,6 @@
 package com.food.common.user.domain;
 
-import com.food.common.common.domain.BaseTimeEntity;
+import com.food.common.basetime.BaseTimeEntity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.Comment;
