+++ user/src/main/java/com/food/user/presentation/AccountController.java
@@ -2,7 +2,7 @@ package com.food.user.presentation;
 
 import com.food.common.apiResult.ApiResult;
 import com.food.common.apiResult.SuccessResult;
-import com.food.common.common.api.ApiFor;
+import com.food.common.annotation.ApiFor;
 import com.food.common.user.enumeration.Role;
 import com.food.user.service.DefaultAccountFindService;
 import com.food.common.user.business.dto.response.accountFind.AccountFindResponse;
